Separate solder paste mask generation into new module
