+++ lib/index.ts
@@ -1,4 +1,5 @@
 export * from "./pcb/convert-circuit-json-to-pcb-svg"
 export * from "./assembly/convert-circuit-json-to-assembly-svg"
 export * from "./sch/convert-circuit-json-to-schematic-svg"
+export * from "./pcb/convert-circuit-json-to-solder-paste-mask"
 export type { ColorMap } from "./utils/colors"
